+++ examples/secure_message_client_encrypt.rs
@@ -37,7 +37,8 @@ fn main() {
         (@arg secret: --secret [path] "Secret key file (default: secret.key)")
         (@arg public: --public [path] "Public key file (default: public.key)")
         (@arg address: -c --connect [addr] "Relay server address (default: localhost:7573)")
-    ).get_matches();
+    )
+    .get_matches();
 
     let secret_path = matches.value_of("secret").unwrap_or("secret.key");
     let public_path = matches.value_of("public").unwrap_or("public.key");