+++ examples/secure_message_server.rs
@@ -29,7 +29,8 @@ fn main() {
         (version: env!("CARGO_PKG_VERSION"))
         (about: "Relay server for Secure Message chat client.")
         (@arg port: -p --port [number] "Listening port (default: 7573)")
-    ).get_matches();
+    )
+    .get_matches();
 
     let port = matches.value_of("port").unwrap_or("7573").parse().unwrap();
     let listen_addr = SocketAddr::new([0; 16].into(), port);